+++ NEWS.md
@@ -42,6 +42,7 @@
 - Derive the type of the Identity resource based on the prefix value of 010 / 001 [MODLD-939](https://folio-org.atlassian.net/browse/MODLD-939)
 - Set LINK of the identity resource based on FOLIO configuration [MODLD-923](https://folio-org.atlassian.net/browse/MODLD-923)
 - Add LIGHT_RESOURCE type to light resources created from MARC 490, 775 & 776 [MODLD-980](https://folio-org.atlassian.net/browse/MODLD-980)
+- Update LabelService to use LabelGeneratorService [MODLD-913](https://folio-org.atlassian.net/browse/MODLD-913)
 
 ## 1.0.3 (04-09-2025)
 - Geographic coverage: single "geographicCoverage" edge is created for Work which contains 043 with multiple $a fields [MODLD-694](https://folio-org.atlassian.net/browse/MODLD-694)